requirements.txt
flask
pandas
langchain
langchain-openai
langchain-community
faiss-cpu
python-dotenv
tiktoken
openai
groq
